extract common presenter methods into module
