+++ app/presenters/image_work_show_presenter.rb
@@ -1,4 +1,6 @@
-class ImageWorkShowPresenter < GeoConcernsShowPresenter
+class ImageWorkShowPresenter < GeoConcerns::GeoConcernsShowPresenter
+  include PlumAttributes
+
   self.work_presenter_class = ::RasterWorkShowPresenter
   self.file_format_service = GeoConcerns::ImageFormatService
 end
