+++ app/presenters/multi_volume_work_show_presenter.rb
@@ -1,4 +1,6 @@
 class MultiVolumeWorkShowPresenter < CurationConcernsShowPresenter
+  include PlumAttributes
+
   def viewing_hint
     'multi-part'
   end
