+++ app/presenters/raster_work_show_presenter.rb
@@ -1,4 +1,6 @@
-class RasterWorkShowPresenter < GeoConcernsShowPresenter
+class RasterWorkShowPresenter < GeoConcerns::GeoConcernsShowPresenter
+  include PlumAttributes
+
   self.work_presenter_class = ::VectorWorkShowPresenter
   self.file_format_service = GeoConcerns::RasterFormatService
 end
