+++ app/presenters/scanned_resource_show_presenter.rb
@@ -1,4 +1,6 @@
 class ScannedResourceShowPresenter < CurationConcernsShowPresenter
+  include PlumAttributes
+
   delegate :has?, :first, to: :solr_document
   self.file_presenter_class = ::FileSetPresenter
   self.work_presenter_class = ::MultiVolumeWorkShowPresenter.work_presenter_class
