+++ app/presenters/vector_work_show_presenter.rb
@@ -1,3 +1,5 @@
-class VectorWorkShowPresenter < GeoConcernsShowPresenter
+class VectorWorkShowPresenter < GeoConcerns::GeoConcernsShowPresenter
+  include PlumAttributes
+
   self.file_format_service = GeoConcerns::VectorFormatService
 end
